Errors Resolved, EDA Checked, File Structure synced
