+++ .gitignore
@@ -119,6 +119,7 @@ venv.bak/
 
 # mkdocs documentation
 /site
+/data
 
 # mypy
 .mypy_cache/
