+++ src/requirements.txt
@@ -1,0 +1,30 @@
+contourpy==1.0.7
+convertdate==2.4.0
+cycler==0.11.0
+fonttools==4.39.0
+hijri-converter==2.2.4
+holidays==0.21
+IProgress==0.4
+ipywidgets==8.0.4
+joblib==1.2.0
+jupyterlab-widgets==3.0.5
+kaggle==1.5.13
+kiwisolver==1.4.4
+korean-lunar-calendar==0.3.1
+lightgbm==3.3.5
+matplotlib==3.7.1
+numpy==1.24.2
+pandas==1.5.3
+Pillow==9.4.0
+pyarrow==11.0.0
+PyMeeus==0.5.12
+pyparsing==3.0.9
+python-slugify==8.0.1
+pytz==2022.7.1
+scikit-learn==1.2.2
+scipy==1.10.1
+seaborn==0.12.2
+text-unidecode==1.3
+threadpoolctl==3.1.0
+tqdm==4.65.0
+widgetsnbextension==4.0.5
